backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
celery
redis
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic
pydantic-settings
pandas
openpyxl
pytest
pgvector
sentence-transformers
requests
oracledb
langchain-ollama
langchain-core
reportlab
python-dateutil